sandbox/csr tidy-up
