+++ sandbox/csr/src/App.tsx
@@ -1,48 +1,10 @@
-import { useSignal, useComputed, For, Derive } from "kaioken"
 import { Router, Route, Link } from "kaioken/router"
 import { ROUTES } from "./routes"
-
-function ForExample() {
-  const items = useSignal([0, 1, 2, 3, 4])
-  const doubledItems = useComputed(() => items.value.map((i) => i * 2))
-  return (
-    <>
-      <ul>
-        <For each={doubledItems}>{(item) => <li>{item}</li>}</For>
-      </ul>
-      <button
-        onclick={() => (items.value = [...items.value, items.value.length])}
-      >
-        Add
-      </button>
-    </>
-  )
-}
-
-function DeriveExample() {
-  const name = useSignal("bob")
-  const age = useSignal(42)
-  const person = useComputed(() => ({ name: name.value, age: age.value }))
-  return (
-    <div>
-      <input bind:value={name} />
-      <input type="number" bind:value={age} />
-      <Derive from={person}>
-        {(person) => (
-          <div>
-            {person.name} is {person.age} years old
-          </div>
-        )}
-      </Derive>
-    </div>
-  )
-}
 
 const Home: Kaioken.FC = () => {
   return (
     <div>
-      <ForExample />
-      <DeriveExample />
+      <h1>Home</h1>
     </div>
   )
 }
